Move interface to its own file
